+++ tests/field/test_core.py
@@ -41,7 +41,8 @@ class TestField:
 
     def test___repr__(self):
         field = Field(rename='test', required=False)
-        assert repr(field) == "Field(default=None, rename='test', required=False, validators=[])"
+        assert repr(field) == ("Field(default=None, deserializers=[], rename='test'"
+                               ', required=False, serializers=[], validators=[])')
 
     def test__setattr__(self):
         field = Field()